+++ meteora/src/cli.rs
@@ -1,5 +1,6 @@
 pub mod delete;
 pub mod get;
+pub mod leave;
 pub mod put;
 pub mod start;
 pub mod status;