+++ CHANGES.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+- Add leave command #10 @mosuka
+- Add status command #9 @mosuka
 - Fix bug for specifying node ID #8 @mosuka
 - Update protobuf #7 @mosuka
 - Refactor protobuf #6 @mosuka